+++ src/plugin.cpp
@@ -57,6 +57,7 @@ void init(Plugin* p) {
 	p->addModel(modelSickoQuant4);
 	p->addModel(modelSickoSampler);
 	p->addModel(modelSickoSampler2);
+	p->addModel(modelSimpleSeq4);
 	p->addModel(modelSlewer);
 	p->addModel(modelSwitcher);
 	p->addModel(modelSwitcherSt);