+++ src/plugin.hpp
@@ -57,6 +57,7 @@ extern Model* modelSickoQuant;
 extern Model* modelSickoQuant4;
 extern Model* modelSickoSampler;
 extern Model* modelSickoSampler2;
+extern Model* modelSimpleSeq4;
 extern Model* modelSlewer;
 extern Model* modelSwitcher;
 extern Model* modelSwitcherSt;