+++ changelog.md
@@ -1,3 +1,12 @@
+### 2.6.17 (2025-01-21)
+- added 'simpleSeq4' module
+- added 'Attenuator' option on adder8 module
+- 'drumPlayerXtra': added 'Randomize samples' in context menu when setting a root folder
+- 'sickoLooper' and 'clocker' modules: improved audio click management
+- 'clocker' and 'clocker2' modules: fixed a bug on /2 clock division that was actually x2, and fixed its display color to red instead of green
+- 'sickoQuant' and 'sickoQuant4' module browser search made easier with keyword 'squant'
+-  multiRouter / multiSwitcher: added 'cycle' and 'RST input = reverse advance' options in the right-click menu
+
 ### 2.6.16 (2024-10-17)
 - added 'keySampler' module.
 - added 'x48' and '/48' multipliyng options on 'clocker' and 'clocker2' modules